Fix for GH Pages not supporting the split filter.
GH Pages runs Liquid v. 2.2.0, but the Split filter (which is what we need to support `<!-- more -->` links), wasn't introduced until 2.3.0.

When GH upgrades, we should use split (code documented inline in `page-single.html`), but for now, the somewhat hacky workaround is to replace <!-- more --> with a `<div>` tag, and hiding the content. Not ideal, especially in terms of SEO, but it should work for now.